cruise ship load for type 1 and type 2
The load has now been simulated for an entire year, the proportion for the variation in the load for the year matches what we know for the month arrivals. The actual value in megawatts is still too high, but if it is the pattern that we want, it is definitely there.  
/A