Add files via upload
This is the first draft of the introduction. I have highlighted areas in yellow,  that I want to remove to make space. In addition, I want to state what the installed capacities that will be used in the model as a part of the background/introduction. Please, this is not perfect, it is just a start, let me know what you think.